Revert "GH-117108: Change the size of the GC increment to about 1% of the total heap size. (GH-117120)"
This reverts commit e28477f214276db941e715eebc8cdfb96c1207d9.

+++ Python/gc_free_threading.c
@@ -1377,7 +1377,7 @@ visit_get_objects(const mi_heap_t *heap, const mi_heap_area_t *area,
 }
 
 PyObject *
-_PyGC_GetObjects(PyInterpreterState *interp, int generation)
+_PyGC_GetObjects(PyInterpreterState *interp, Py_ssize_t generation)
 {
     PyObject *result = PyList_New(0);
     if (!result) {
